backend/requirements.txt
# pip install -r requirements.txt

google-genai==1.47.0
boto3==1.41.2
requests==2.32.5
Pydantic==2.12.4
pytz==2025.2
python-dotenv==1.2.1
